request

+++ app/src/main/java/it/polito/kgame/ui/calendar/CalendarFragment.kt
@@ -43,17 +43,19 @@ class CalendarFragment : Fragment(R.layout.fragment_calendar) {
 
         val mImpegniDays: MutableList<EventDay> = ArrayList()
         val listaimpegni : MutableList<EventoInfo> = ArrayList()
-
+        mImpegniDays.clear()
+        listaimpegni.clear()
         rvdettagli.isVisible = false
         ante.isVisible=false
         ante.y = 0F
-
+        calendarView.setEvents(mImpegniDays)
         println("top: $ante.y")
         calendarViewModel.Appointments.observe(viewLifecycleOwner, Observer { appointment ->
             println("APPOINTMENTS: $appointment")
             println("napp: ${appointment.size}")
             mImpegniDays.clear()
             listaimpegni.clear()
+            calendarView.setEvents(mImpegniDays)
             appointment.forEach {
                 println("itInizio: $it")
                 var x = java.util.Calendar.getInstance()
@@ -237,10 +239,31 @@ class CalendarFragment : Fragment(R.layout.fragment_calendar) {
                     close.setOnClickListener {
                         ante.isVisible = false
                     }
+
                     delete.setOnClickListener {
-                        DbManager.deleteAppointment(I!!)
-                        adapter.deleteEvento(I!!)
-                        ante.isVisible = false
+
+                        MaterialAlertDialogBuilder(requireContext())
+                                .setTitle("Elimina impegno")
+                                .setMessage("Sei sicuro di voler eliminare questo impegno?")
+                                .setPositiveButton("Si") { _, _ ->
+                                    ante.isVisible = false
+                                    DbManager.deleteAppointment(I!!)
+                                    adapter.deleteEvento(I!!)
+
+
+                                    Toast.makeText(
+                                            requireContext(), "Impegno eliminato correttamente ",
+                                            Toast.LENGTH_SHORT
+                                    ).show()
+
+
+
+                                }
+                                .setNegativeButton("No"){  _, _ ->
+
+                                }
+                                .show()
+
                     }
                     buttonevent.setOnClickListener {
                         ante.isVisible = false
@@ -267,10 +290,7 @@ class CalendarFragment : Fragment(R.layout.fragment_calendar) {
                         }
 
                     }
-                    /*deletedettagli.setOnClickListener {
-                        DbManager.deleteAppointment(impegnidatacorrente[0])
-                        adapter.deleteEvento(impegnidatacorrente[0])
-                    }*/
+
                 }
                 else {
                     Log.d("tag", "Non ha impegno")
